update pipelines folder structure and README
